+++ docs/changelog.md
@@ -10,11 +10,13 @@ This project uses semantic versioning and follows [keep a changelog](https://kee
 - Drop support for python 3.9
 - Support for python 3.14
 - Accept `pathlib.Path`s in `get_evaluable_architecture(..)`
+- Readme examples and links to documentation
 
 ### Fixed
 - Straightforward error message when using wildcards in `are_named` rules.
 - Typo in docstring of `get_evaluable_architecture(..)`.
 - Module filter uses regex when activated instead of only if last module.
+- Descriptive error message with actionable guidance when a queried module is not present in the dependency graph.
 
 ## 4.0.1 -- 2025-08-08
 ### Fixed